refactor(common): move the `geospatial` module to the base SQL backend
it contains routines to compile geospatial values as SQL

BREAKING CHANGE: removed `ibis.common.geospatial`, import the functions from `ibis.backends.base.sql.registry.geospatial`
