+++ ibis/backends/base/sql/registry/geospatial.py
